Added a scraper for datacite records.
This is a longish script and could use some clean-up subsequently, but the script uses the pytacite package to pull in datacite objects minted using Zenodo that are software hosted on Github. As much as possible it then adds the repository and pulls in the appropriate datacite metadata so we can do additional work.